Feat: 필터링 기능 수정 - 날짜순 정렬시 주간 일간순 정렬과 겹치는 문제 수정 Github issue #8
