+++ src/app/page.tsx
@@ -4,8 +4,10 @@ import { HomeSkeleton } from './components/homeSkeleton';
 import { fetchYoutubeChannels } from '@/actions/youtube/youtubeThumbnails.action';
 import { Navbar } from '@/components/common/navBar';
 import { getVideos } from '@/actions/youtube/video.action.ts';
+import { fetchYoutubeVideos } from '@/actions/youtube/youtube.action';
 
 export default async function PlaylistViewer() {
+  await fetchYoutubeVideos();
   const [latestVideos, topVideo, initialChannel] = await Promise.all([
     getVideos({ sortBy: 'date' }),
     getVideos({ sortBy: 'views', limit: 5 }),
@@ -15,6 +17,7 @@ export default async function PlaylistViewer() {
   return (
     <Suspense fallback={<HomeSkeleton />}>
       <Navbar />
+
       <UI latestVideos={latestVideos.videos} topVideo={topVideo.videos} initialChannel={initialChannel} />
     </Suspense>
   );
