+++ src/hooks/useVideoFilters.ts
@@ -13,12 +13,20 @@ export function useVideoFilters(initialFilters: VideoFilters) {
     }),
     [searchParams, initialFilters]
   );
-
   const updateFilter = useCallback(
     (filterType: keyof VideoFilters, newValue: string) => {
       if (filters[filterType] === newValue) return;
 
       const params = new URLSearchParams(searchParams);
+      if (filters['sort'] === 'date' && filterType === 'rankType' && newValue !== 'total') {
+        params.set('sort', 'views');
+      }
+      if (filterType === 'sort' && newValue === 'date') {
+        const currentRankType = params.get('rankType');
+        if (['daily', 'weekly'].includes(currentRankType)) {
+          params.set('rankType', 'total');
+        }
+      }
       params.set(filterType, newValue);
       router.replace(`?${params.toString()}`);
     },
